feat(record): add Playwright and YAML test generation utilities with shared types and export controls
