+++ apps/chrome-extension/src/extension/record/generators/index.ts
@@ -1,0 +1,7 @@
+// Main generators
+export { generatePlaywrightTest } from './playwrightGenerator';
+export { generateYamlTest, exportEventsToYaml } from './yamlGenerator';
+
+// Shared utilities
+export * from './shared/types';
+export * from './shared/testGenerationUtils';
